+++ ionic2/app/components/filter-popover/filter-popover.ts
